Add dry run option to snis_update_assets script

+++ util/snis_update_assets.sh
@@ -8,6 +8,7 @@ up_to_date_count=0
 new_count=0
 update_count=0
 update_fail_count=0
+export dryrun=0
 
 sanity_check_environment()
 {
@@ -21,16 +22,20 @@ sanity_check_environment()
 
 fetch_file()
 {
-	URL="$1"
-	FILE="$2"
+	local URL="$1"
+	local FILE="$2"
+	local dryrun="$3"
 	echo -n 1>&2 "Fetching $URL... "
-	wget --quiet "$1" -O - > "$FILE"
-	if [ "$?" != "0" ]
+	if [ "$dryrun" = "0" ]
 	then
-		/bin/rm -f "$FILE"
-		echo 1>&2
-		echo "$PROG"': Failed to fetch '"$URL" 1>&2
-		return 1
+		wget --quiet "$1" -O - > "$FILE"
+		if [ "$?" != "0" ]
+		then
+			/bin/rm -f "$FILE"
+			echo 1>&2
+			echo "$PROG"': Failed to fetch '"$URL" 1>&2
+			return 1
+		fi
 	fi
 	echo "done" 1>&2
 	return 0
@@ -38,16 +43,20 @@ fetch_file()
 
 move_file()
 {
-	old="$1"
-	new="$2"
-	mv -f "$old" "$new"
+	local dryrun= "$1"
+	local old="$2"
+	local new="$3"
+	if [ "$dryrun" = "0" ]
+	then
+		mv -f "$old" "$new"
+	fi
 	return $?
 }
 
 update_file()
 {
-	checksum="$1"
-	filename="$2"
+	local checksum="$1"
+	local filename="$2"
 	if [ -f "$filename" ]
 	then
 		localchksum=$(md5sum $filename | awk '{ print $1 }')
@@ -55,13 +64,13 @@ update_file()
 		then
 			up_to_date_count=$((up_to_date_count + 1))
 		else
-			move_file "$filename" "$filename".old
+			move_file "$dryrun" "$filename" "$filename".old
 			if [ "$?" != "0" ]
 			then
 				echo "$PROG"':Cannot move old $filename out of the way, skipping' 1>&2
 				update_fail_count=$((update_fail_count + 1))
 			else
-				fetch_file "$ASSET_URL"/"$filename" "$filename"
+				fetch_file "$ASSET_URL"/"$filename" "$filename" "$dryrun"
 				if [ "$?" != "0" ]
 				then
 					update_fail_count=$((update_fail_count + 1))
@@ -71,19 +80,22 @@ update_file()
 			fi
 		fi
 	else
-		dname=$(dirname "$filename")
+		local dname=$(dirname "$filename")
 		if [ ! -d "$dname" ]
 		then
-			mkdir -p $dname
-			if [ "$?" != "0" ]
+			if [ "$dryrun" = "0" ]
 			then
-				echo "$PROG"': Failed to create directory for '"$filename" 1>&2
-				update_fail_count=$((update_fail_count + 1))
+				mkdir -p $dname
+				if [ "$?" != "0" ]
+				then
+					echo "$PROG"': Failed to create directory for '"$filename" 1>&2
+					update_fail_count=$((update_fail_count + 1))
+				fi
 			fi
 		fi
-		if [ -d "$dname" ]
+		if [ -d "$dname" -o "$dryrun" != "0" ]
 		then
-			fetch_file $ASSET_URL/$filename $filename
+			fetch_file $ASSET_URL/$filename $filename "$dryrun"
 			if [ "$?" = "0" ]
 			then
 				new_count=$((new_count+1))
@@ -110,16 +122,31 @@ update_files()
 
 output_statistics()
 {
+	if [ "$dryrun" = "0" ]
+	then
+		updated="updated"
+		new="new files"
+	else
+		updated="would be updated"
+		new="new files would be created"
+	fi
+
 	echo
 	echo "$up_to_date_count files already up to date"
-	echo "$update_count files updated"
-	echo "$new_count new files"
+	echo "$update_count files $updated"
+	echo "$new_count $new"
 	echo "$update_fail_count update failures"
 	echo
 }
 
+if [ "$1" = "--dry-run" ]
+then
+	dryrun=1
+	shift;
+fi
+
 sanity_check_environment
-fetch_file "$MANIFEST_URL" "$MANIFEST_FILE"
+fetch_file "$MANIFEST_URL" "$MANIFEST_FILE" 0
 if [ "$?" != "0" ]
 then
 	exit 1
